Moved minifyCriteria to js module
